refactor: split server.py monolith into 12 domain router modules (#120)
server.py: 2652 LOC → 196 LOC (bootstrap + backward-compat re-exports).
60 @mcp.tool() handlers split into domain modules under routers/:

  books.py (6), chapters.py (8), authors.py (4), state.py (7),
  creation.py (3), memoir.py (2), series.py (3), ideas.py (5),
  gates.py (8), reference.py (4), claudemd.py (8), scenes.py (2)

All routers import the shared mcp + _cache from routers/_app.py.
server.py re-exports every tool function for backward compat and
uses a PatchMirroringModule shim so legacy monkeypatch.setattr(server,
"load_config", ...) tests keep working without modification.

All 1070 tests pass.
