+++ servers/storyforge-server/routers/authors.py
@@ -1,0 +1,140 @@
+"""Author-profile CRUD tools: list/get/create/update."""
+
+from __future__ import annotations
+
+import json
+from datetime import date
+
+import yaml
+
+from tools.shared.paths import resolve_author_path, slugify
+from tools.state.parsers import parse_frontmatter
+
+from . import _app
+from ._app import _cache, mcp
+
+
+@mcp.tool()
+def list_authors() -> str:
+    """List all author profiles."""
+    state = _cache.get()
+    authors = state.get("authors", {})
+    result = [
+        {"slug": slug, "name": a.get("name", slug),
+         "genres": a.get("primary_genres", []),
+         "studied_works": a.get("studied_works_count", 0)}
+        for slug, a in authors.items()
+    ]
+    return json.dumps({"authors": result, "count": len(result)})
+
+
+@mcp.tool()
+def get_author(slug: str) -> str:
+    """Get full author profile data."""
+    state = _cache.get()
+    author = state.get("authors", {}).get(slug)
+    if not author:
+        return json.dumps({"error": f"Author '{slug}' not found"})
+
+    # Also load vocabulary if exists
+    config = _app.load_config()
+    vocab_path = resolve_author_path(config, slug) / "vocabulary.md"
+    if vocab_path.exists():
+        author["vocabulary"] = vocab_path.read_text(encoding="utf-8")
+
+    return json.dumps(author)
+
+
+@mcp.tool()
+def create_author(name: str, genres: str = "", tone: str = "", voice: str = "third-limited", tense: str = "past") -> str:
+    """Create a new author profile directory with template files.
+
+    Args:
+        name: Author pen name
+        genres: Comma-separated primary genres
+        tone: Comma-separated tone descriptors (e.g. "sarcastic, dark-humor")
+        voice: Narrative voice (first-person, third-limited, third-omniscient, second-person)
+        tense: Narrative tense (past, present)
+    """
+    config = _app.load_config()
+    slug = slugify(name)
+    author_dir = resolve_author_path(config, slug)
+
+    if author_dir.exists():
+        return json.dumps({"error": f"Author '{slug}' already exists"})
+
+    author_dir.mkdir(parents=True)
+    (author_dir / "studied-works").mkdir()
+    (author_dir / "examples").mkdir()
+
+    genre_list = [g.strip() for g in genres.split(",") if g.strip()] if genres else []
+    tone_list = [t.strip() for t in tone.split(",") if t.strip()] if tone else []
+
+    today = date.today().isoformat()
+    profile = f"""---
+name: "{name}"
+slug: "{slug}"
+created: "{today}"
+updated: "{today}"
+primary_genres: {json.dumps(genre_list)}
+narrative_voice: "{voice}"
+tense: "{tense}"
+tone: {json.dumps(tone_list)}
+sentence_style: "varied"
+vocabulary_level: "moderate"
+dialog_style: "naturalistic"
+pacing: "tension-driven"
+themes: []
+influences: []
+avoid: ["purple-prose", "info-dumps", "deus-ex-machina"]
+---
+
+# {name}
+
+## Writing Style
+
+*Style description will be refined through the study-author skill.*
+
+## Signature Techniques
+
+- *To be defined*
+
+## Voice Notes
+
+*Notes on this author's distinctive voice characteristics.*
+"""
+
+    (author_dir / "profile.md").write_text(profile, encoding="utf-8")
+    (author_dir / "vocabulary.md").write_text(
+        f"# {name} — Vocabulary\n\n## Preferred Words\n\n*To be defined*\n\n## Banned Words\n\n- delve\n- tapestry\n- nuanced\n- vibrant\n- resonate\n- pivotal\n- multifaceted\n- realm\n- testament\n- intricate\n- myriad\n- unprecedented\n- foster\n- beacon\n- juxtaposition\n- paradigm\n- synergy\n- interplay\n- ever-evolving\n- navigate (metaphorical)\n\n## Signature Phrases\n\n*To be defined through study-author*\n",
+        encoding="utf-8",
+    )
+
+    _cache.invalidate()
+    return json.dumps({
+        "success": True,
+        "slug": slug,
+        "path": str(author_dir),
+        "message": f"Author profile '{name}' created at {author_dir}",
+    })
+
+
+@mcp.tool()
+def update_author(slug: str, field: str, value: str) -> str:
+    """Update a field in an author's profile frontmatter."""
+    config = _app.load_config()
+    profile_path = resolve_author_path(config, slug) / "profile.md"
+
+    if not profile_path.exists():
+        return json.dumps({"error": f"Author '{slug}' not found"})
+
+    text = profile_path.read_text(encoding="utf-8")
+    meta, body = parse_frontmatter(text)
+    meta[field] = value
+    meta["updated"] = date.today().isoformat()
+
+    new_text = "---\n" + yaml.dump(meta, default_flow_style=False, allow_unicode=True) + "---\n" + body
+    profile_path.write_text(new_text, encoding="utf-8")
+
+    _cache.invalidate()
+    return json.dumps({"success": True, "field": field, "value": value})
