+++ servers/storyforge-server/routers/scenes.py
@@ -1,0 +1,154 @@
+"""Snowflake Method scene-list tools.
+
+``plot/scenes.md`` is a structured Markdown table that lets the writer
+plan beats at scene granularity (Snowflake Step 8). The two tools here
+create and incrementally update individual rows.
+"""
+
+from __future__ import annotations
+
+import json
+
+from tools.shared.paths import resolve_project_path
+from tools.state.parsers import parse_frontmatter
+
+from . import _app
+from ._app import _cache, mcp
+
+
+@mcp.tool()
+def create_scene_list(
+    book_slug: str,
+    scenes: list[dict],
+) -> str:
+    """Create or overwrite plot/scenes.md with a scene list (Snowflake Step 8).
+
+    Each scene dict may contain:
+      - number (int, required): sequential scene number
+      - chapter (str): e.g. "Ch. 01"
+      - pov (str): POV character name
+      - summary (str): one-sentence scene description
+      - est_words (int): estimated word count
+      - status (str): Planned / Written / Revised / Final
+
+    Args:
+        book_slug: Book project slug
+        scenes: List of scene dicts describing each scene
+    """
+    config = _app.load_config()
+    project_dir = resolve_project_path(config, book_slug)
+    if not project_dir.exists():
+        return json.dumps({"error": f"Book '{book_slug}' not found"})
+
+    scenes_path = project_dir / "plot" / "scenes.md"
+
+    title = book_slug.replace("-", " ").title()
+    readme = project_dir / "README.md"
+    if readme.exists():
+        fm, _ = parse_frontmatter(readme.read_text(encoding="utf-8"))
+        title = fm.get("title", title)
+
+    header = (
+        f"# Scene List: {title}\n\n"
+        "*Generated by Snowflake Method — Step 8. "
+        "Update status via MCP `update_scene()` as you write.*\n\n"
+        "## Scene Index\n\n"
+        "| # | Chapter | POV | Scene Summary | Est. Words | Status |\n"
+        "|---|---------|-----|---------------|------------|--------|\n"
+    )
+    rows = []
+    for s in scenes:
+        num = s.get("number", "?")
+        chapter = s.get("chapter", "")
+        pov = s.get("pov", "")
+        summary = s.get("summary", "")
+        est_words = s.get("est_words", "")
+        status = s.get("status", "Planned")
+        rows.append(f"| {num} | {chapter} | {pov} | {summary} | {est_words} | {status} |")
+
+    footer = (
+        "\n## Status Key\n\n"
+        "- **Planned** — Scene defined, not yet written\n"
+        "- **Written** — First draft complete\n"
+        "- **Revised** — At least one revision pass done\n"
+        "- **Final** — Scene is locked\n\n"
+        "## Notes\n\n"
+        "*Cross-scene notes: recurring motifs, foreshadowing chains, POV balance.*\n"
+    )
+
+    content = header + "\n".join(rows) + "\n" + footer
+    scenes_path.write_text(content, encoding="utf-8")
+    _cache.invalidate()
+
+    return json.dumps({
+        "success": True,
+        "path": str(scenes_path),
+        "scene_count": len(scenes),
+        "message": f"Scene list created with {len(scenes)} scenes at {scenes_path}",
+    })
+
+
+@mcp.tool()
+def update_scene(
+    book_slug: str,
+    scene_number: int,
+    chapter: str = "",
+    pov: str = "",
+    summary: str = "",
+    est_words: str = "",
+    status: str = "",
+) -> str:
+    """Update one scene row in plot/scenes.md (Snowflake Method).
+
+    Only pass the fields you want to change — omitted fields are preserved.
+
+    Args:
+        book_slug: Book project slug
+        scene_number: The scene's sequential number (# column)
+        chapter: Chapter assignment, e.g. "Ch. 05"
+        pov: POV character name
+        summary: One-sentence scene summary
+        est_words: Estimated word count (string, e.g. "1200")
+        status: Planned / Written / Revised / Final
+    """
+    config = _app.load_config()
+    project_dir = resolve_project_path(config, book_slug)
+    scenes_path = project_dir / "plot" / "scenes.md"
+
+    if not scenes_path.exists():
+        return json.dumps({"error": f"scenes.md not found for '{book_slug}'. Run create_scene_list first."})
+
+    lines = scenes_path.read_text(encoding="utf-8").splitlines(keepends=True)
+    updated = False
+
+    for i, line in enumerate(lines):
+        if not line.startswith("|"):
+            continue
+        cells = [c.strip() for c in line.strip().strip("|").split("|")]
+        if len(cells) < 6:
+            continue
+        if cells[0] == str(scene_number):
+            new_chapter = chapter if chapter else cells[1]
+            new_pov = pov if pov else cells[2]
+            new_summary = summary if summary else cells[3]
+            new_est_words = est_words if est_words else cells[4]
+            new_status = status if status else cells[5]
+            lines[i] = (
+                f"| {scene_number} | {new_chapter} | {new_pov} | "
+                f"{new_summary} | {new_est_words} | {new_status} |\n"
+            )
+            updated = True
+            break
+
+    if not updated:
+        return json.dumps({"error": f"Scene #{scene_number} not found in scenes.md"})
+
+    scenes_path.write_text("".join(lines), encoding="utf-8")
+    _cache.invalidate()
+
+    return json.dumps({
+        "success": True,
+        "scene_number": scene_number,
+        "path": str(scenes_path),
+        "message": f"Scene #{scene_number} updated",
+    })
